src/requirements.txt
cv2_enumerate_cameras==1.1.18.2
#imutils==0.5.4
lmfit==1.3.2
matplotlib
mediapipe==0.10.21
numpy
opencv_contrib_python==4.11.0.86
opencv_python==4.10.0.84
pandas==2.2.3
perlin_numpy==0.0.1
Pillow==11.1.0
pygame==2.6.1
scikit_learn==1.6.1
scipy
simple_pid==2.0.1
scikit-image
ultralytics==8.3.59
lap>=0.5.12